qdoc3: Updating qdoc's QML parser (preparatory work).
For documenting groups of QML properties when documenting
a QML element in a C++ file, the concept of a QML property
group was added, which created an intermediate node in the
qdoc node tree to represent a group of QML properties. This
was requested so that groups of QML properties could
documented with a single qdoc comment. But for documenting
QML components directly in .qml files, the property group
is not useful, so the HTML and DITA XML output generators
need to know they can see trees that don't contain the QML
property group nodes above the QML property nodes. This is
the change that fixes that.

But there is also added here some infrastructure for
processing qdoc metacommands in the QML parser's vistor
functions.

Task-number:  QTBUG-21229
Change-Id: Ic0f6cafbb098132c814d39067dc71d98b61edfce
Reviewed-on: http://codereview.qt-project.org/4417
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>