qdoc3: Change some headers in the docs
There are some headings on C++ and QML reference pages that concern
both signals and signal handlers. Currently, the headings include the
phrase "Signal Handlers," although many times what is documented there
are signals, not signal handlers. Sometimes, however, it is signal
handlers that get documented. The problem is we don't all the
documenter to specify which one he is documenting. That's the real
problem, and it requires adding a \qmlsignalhandler command to qdoc,
or something like that. But for now, this task just changes the
headings in the documentation back to "Signals" because that is a bit
less confusing.

Task-number:  QTBUG-21542
Change-Id: I824821f8d233163ed1d9513c24d5122e13331a57
Reviewed-on: http://codereview.qt-project.org/5216
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>