qdoc3: Fixes the real problem for QTBUG-21542
Because we can now document QML components in .qml files,
we now have the opportunity to document signal handlers.
A signal handler is a code block that binds some javascript
to a label of the form "onXxx:" where Xxx is the name of a
signal. You can now place a qdoc comment directly above the
signal handler label. The comment has no topic command in it.
It is just a paragraph (or more) stating that the component
handles the Xxx signal and this is what it does, etc.

Task-number:  QTBUG-21542
Change-Id: I3bd7e65661bc5f3ef10eac9ad0ec5a04961b1230
Reviewed-on: http://codereview.qt-project.org/5379
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>