qdoc3: qdoc now recognizes \qmlabstract
It doesn't do anything with it yet, but in the
course of doing this first bit, I saw that the
QML parser was missing some code to recognize
the \ingroup command and the \inqmlmodule
command, and a few others. So it is important
to push this part now.

Task-number:  QTBUG-22557
Change-Id: I3ee85547d2feb2dcbfa1b4d4160d7b0e50e0bacd
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>