qdoc3: qdoc now reads qt.index better
It was creating the wrong kind of node when it saw
a qmlclass or a qmlbasictype. Now it creates the
correct node, but there could still be incorrect
behavior here, so this investigation is ongoing.
But the bug that caused the crash is fixed.

Task-number:  QTBUG-22462
Change-Id: Ib80129fdd9fd63a60c0022c78c2bbd036946f8a6
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>