qdoc3: qdoc suffered a regression bug
qdoc had stopped listing properties, signals, and
methods for QML elements. This regression was probably
caused by changes made for the purpose of allowing
qdoc to let any doc page be included in a QML module.
This resulted the module qualifiers being stored as
the keys in a map, but subsequent lookups in the map
used the module identifier. They are slightly different,
although we might not really need both of them in the
long run. In this case, we want to use the module
identifier in both places. Fixed.

Task-number:  QTBUG-22684
Change-Id: I6cf057f86319f5d279f0b3d5e80233dc1d8b7d5c
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>