qdoc3: qdoc now uses \qmlabstract
Suppose you have created a QML component
that you want other QML components to be
based on. Other QML components will inherit
this component, but you don't intend that
this component will be used on its own.
Then in the \qmlclass comment for that
component, you include \qmlabstract.
No argument follows the command.

Now all properties, signals, and
methods available in the abstract
component will be listed on the
reference page for each QML component
that inherits the abstracft component,
as if they were defined in the
inheriting components.

This fix also removes the
qmlModuleQualifier() function and
replaces all calls to it with
qmlModuleIdentifier().

Task-number:  QTBUG-22557
Change-Id: I71b2860a8d6b4e201ec0c52aa780a78812c88d33
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>