qdoc3: qdoc now has a \readonly command
Currently, there is no way to tell qdoc3 to mark a
property in a QML component as read-only. For QML
elements in QtQuick 1 and QtQuick 2, qdoc3 tries to
figure out if a WML property is read-only or writable,
and it markes the ones it thinks are read-only with a
"read-only" tag in the output.But this doesn't work
very well for QML elements, and it doesn't work at all
for QML components.

So now you can include the \readonly command (no arg)
in the qdoc comment for a \qmlproperty, and qdoc3
will mark that property with "read-only" in the
description of that property in the output.

Note this works for both QML components and elements.
Note to QTQuick 1 & 2 devs: This command works better
than letting qdoc3 try to figure it out

Task-number:  QTBUG-22624
Change-Id: I7b42f392adec483cff9b3c019169333d0b71cceb
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>