qdoc3: Don't output 0 value in hex for DITA XML
The value attribute of the cxxEnumeratorInitialiser tag can be 0x0 or
0x00 etc, but some publishers treat this value as the null
value. Enumerator values appear in hex when they are used for setting
bits on and off. But when the actual value is 0, there is no need to
write the value in hex. The DITA XML generator is changed to output
0 instead of 0x0 for hex 0 values for enumerator initialisers.

Task-number:  QTBUG-23585
Change-Id: I7d4ac1ccfa8493b0db82c08a4236758305d94a4c
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>