qdoc3: The QML parser now reports syntax errors
The QML parser in qtdeclarative is used by qdoc to parser both .qml
files and QML code snippets. Currently, it only reports errors found
when parsing snippets. It should also report errors when parsing QML
component files. If the QML parser detects an error when parsing a QML
component file, the documentation for the component is not
generated. That's the way the parser works. It first parses the file
and builds a parse tree. It only traverses the tree if the parse was
successful.

Task-number:  QTBUG-23762

Change-Id: I02e4cc6e78a3226376f295b27b787786c14f4a28
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>