qdoc3: id attributes in DITA XML now begin with "id-"
It was suggested that every id attribute could begin with "id-" to
prevent all cases of illegal first characters. This suggestion is
adopted. Note that when this change is integrated, all qdoc output
must be rebuilt because existing id attributes will no longer be valid
in links. This problem will only affect those who are depending on id
attributes remaining unchanged from one qdoc run to the next.

Task-number:  QTBUG-24118
Change-Id: Iab2d5f5270aff6001348f3ece47e9fe8a5b518b0
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>