qdoc3: Allow \qmlproperty to override the property signature
Now you can use a list of \qmlproperty commands
in a single qdoc comment, when you are documenting
a QML property in a .qml file. The signature of the
QML property being docunenbted is effectively
overridden in the documentation by the signature(s)
of the \qmlproperty command(s).

Task-number:  QTBUG-23922
Change-Id: I10c5f468d8c1acf027b7cb5b66f5465ec9a4d0cb
Reviewed-by: Martin Smith <martin.smith@nokia.com>
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>