qdoc3: Fix qdoc's use of vague QML terms
qdoc already uses only the QML type name as the title for a QML class
reference page for what used to be called QML elements. Now it should
do the same thing for QML components, i.e. just use the QML class name
as the title of the reference page for a QML component. Also clean up
the breadcrumbs use of Element and Component.

Task-number:  QTBUG-24545
Change-Id: Ie19f3541821337b29fb1b130188a33b3e46ebaf8
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>