Add some of the QML documentation to the doc build.
This change re-adds some of the QML documentation back to the build done
by running the "make docs" command in this module. Not all directories
have been added, since it is not clear if they belong with the QtQml or
QtQuick module. The directories could arbitrarily be added to one of the
qdocconf files, but that might not be the best decision when we want to
modularize the documentation properly.

Change-Id: Ia16f276abf02916ed96cc6e146745187898dee0a
Reviewed-by: Jerome Pasion <jerome.pasion@nokia.com>