Prevent 'make docs' (from qt5.git) from generating monolithic docs
We don't want to recurse into the qtdoc repo and do a full monolithic
doc build as part of a modularized docs build of qt5.git. At some point
we'll have to factor out the global/generic docs out of this repo into
a qglobaldocs repo, or just throw away all the monotholitic stuff and
keep the generic docs here.

Change-Id: I646350ca2dea0756877179b39519616ef7e173ff
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>
Reviewed-by: Martin Smith <martin.smith@digia.com>
Reviewed-by: Johanna Äijälä <johanna.aijala@digia.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>