Make the qtdoc repo part of the modularized build
The monolithic build is now partially removed, but there's still a lot
of cruft left that we'll remove once we know what we need or not need.

This introduces a bunch of qdoc warnings for examplee documentation that
live in the qdoc repo, but that references an example that lives in the
module. These example docs should be moved to the corresponding module.

Change-Id: I15f9f9087f1e96786ce2a49a7192f1015cb192db
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>