Fix broken snippets in Getting Started with QML
Most of the code snippets in this file are inlined, so it makes
no sense to have these three out-of-line. They were also broken
because we don't do cross-repository linking of snippets. At
some point, the docs should be moved into the qtdeclarative
repository, but to have a safe fix for right now, I've just
copy-pasted the snippets from the example files. This makes the
file consistent at least and makes the missing snippets appear
in the  documentation.

Change-Id: I2614eb456107ec0ea334f02d95e8b6af5551b74c
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>