Resolve "Porting QML Applications to Qt 5" name conflict
QtQuick docs also contain a page which should be called "Porting QML
Applications to Qt 5". As that page contains the complete list of
changes which may affect porting, it should have the more generic name.

The page in qtdoc has been rename "Porting QML Applications to Qt 5
Example" as it is a worked example of porting flickr.

Task-number: QTBUG-28411
Change-Id: I2b883a2bfa0abc3a978289a31013aa55e913bce1
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>