Doc: Do not parse snippets as documentation source
The .qdocconf file was using the root of the repository as a source
directory, leading to also the snippets being parsed as source for
docs.

This change modifies .qdocconf to remove the unnecessary paths and
excludes src/snippets. It also fixes the \snippet links to be
relative to src/ instead of the root folder.

Change-Id: Ifd03d7f5f28360bbe89b250fc4567a81ab3a008d
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>