Doc: Move generic code sample image and add it to qhp
In order to add extra images to a qhp file, we need to both
add it to output directory using HTML.extraimages, and from
there to qhp using qhp.QtDoc.extraFiles.

Also move the image to a better location and resize it
to look better in Qt Creator.

Task-number: QTBUG-31314
Change-Id: I817c9bb0e464b1f05430b4191b3bea21961ea122
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>