Doc: Removed a reference to the Qt Creator version
The version information was added at the time when
the Qt installation package didn't ship the creator
version that supported Qt for Android. Now that the
latest Qt installation package has the latest version
of Qt Creator with Qt for Android support, this
information can go.

Besides the above mentioned change, changed the
title of Android platform page to be in sync with the
other platform pages, which follow the "Qt for.."
naming convention.

Task-number: QTBUG-33650
Change-Id: I55d9a10a0f7f95a5bd4a62003a4a875c379a852c
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>