Explain how to identify missing dependencies for Android deployment
When you deploy on Android, the tool will attempt to guess which
plugins you need. This can be confusing to many who are trying
to load e.g. images with formats that require a specific plugin
to load. So we explain in the documentation how you can debug this
and find out what modules to add to your .pro file.

Task-number: QTBUG-34586
Change-Id: Iaa13e0cc4183e3574d1b0d66a1352720c6e45753
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>