Doc: Removed an unnecessary step from the porting instructions
Using the absolute URL to the resource stored in a qrc avoids the need
for the removed step. i.e. The "qrc:" scheme-based URLs
to load a resource from the qrc file must include the PREFIX defined
while creating the qrc file. Without the PREFIX, the relative references
in the QML document will not be resolved, unless they are expcitly
imported with a qualified local namespace.

Task-number: QTBUG-40300
Change-Id: I3f9d02c4c1110c4b8c63b45ad8e54f27007c0218
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>