Doc: Edit section titles to prevent incorrect linking
When resolving a link, QDoc now searches the local module first.
This means that if a local module has section titles that collide
with the titles in other modules, and results in incorrect linking
without any warnings being generated. For example, a link to
'Qt Core' would go to a similarly-named section in this module,
instead of the landing page of the Qt Core module documentation.

This change works around the issue by renaming section titles to
be more specific, and by disambiguating the links with the target
module name on 'All Qt Modules' page.

Task-number: QTBUG-40945
Change-Id: Ib868b98f2de19986857292a242b0d8a2db2f382a
Reviewed-by: Martin Smith <martin.smith@digia.com>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>