Doc: Add sidebar content used in the qt.io online template
As part of the new online style, each page has a sidebar with
links to common topics - the same content as on the Qt reference
documentation (index) page.

Switch the order of the topics on the sidebar, based on Google
Analytics feedback. Most-used links are from the 'Reference'
category, so raise that one to the top.

This commit adds the raw .html file into /style, which is then
loaded dynamically by the template code.

Task-number: QTBUG-42086
Change-Id: I69add8ca75c00bb1ff7039adf34194c2c673885d
Reviewed-by: Martin Smith <martin.smith@digia.com>