Doc: Remove superfluous example
The files of the 'texteditor' example under the /examples are not referred
to in the tutorial documentation that originally introduced it, nor is any
example documentation generated for it.

Remove the example as it ends up being copied into binary packages, which
is not intended.

This commit partially reverts 5cd576ea736d94cf0c95d4325d6367873ad338df.

Change-Id: Iff121f6f0a55c5b47cd6420559e184f599db3a3e
Task-number: QTBUG-48501
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>