Doc: Remove docs for old, deleted, platform-specific functions
Most of these functions were removed prior to the release of Qt 5.0.
Three of them remain in the code base though:

1. qt_set_sequence_auto_mnemonic()
2. qt_mac_set_dock_menu()
3. qt_mac_secure_keyboard()

(1) is currently referenced by the QShortcut class ref, so I'm
transferring its documentation to QKeySequence.

(2) is already documented under QMenu.

(3) is not mentioned anywhere else in the Qt docs, so I'll just leave it
undocumented and let it be quietly replaced with a better function.

Change-Id: I27ecc79de7967593781c5cb641e7ec14752b9a3b
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>