Doc: Update supported platforms information
Remove the separate table of reference configurations, and include
the list of supported platforms/configurations that was previously
maintained separately outside of this repository.

The supported configurations are now listed in a single file,
supported-platforms.qdocinc, and included into the main
'Supported Platforms' page.

Separate sections linking to different platform pages are removed,
and replaced with the links within the supported configurations
table headers.

Change-Id: I991e4ee2c69c3e8992b34289cb869f53521a05be
Reviewed-by: Sami Makkonen <sami.makkonen@theqtcompany.com>
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>