Remove version-specific information from the OS X related doc pages.
This removes several factoids which are actually incorrect now (Qt 5.6
has no support whatsoever for 10.6, and 10.7 is NOT a reference
configuration). Instead of updating version numbers, we remove them
entirely and instead rely on the supported platforms page to inform
users of the available versions for development and deployment. This
ensures that the documentation will be kept in sync with reality.

Change-Id: I3c1807661c1da0517b957f942b34c7ef97549d16
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>