Document BlueZ license dependency in QtBluetooth sdpscanner
For most Qt and BlueZ interaction DBus is used which provides a clean
interface from a license perspective. The only exception is the SDP scan
code which directly interacts via BlueZ's SDP APIs. Sdpscanner
is called via QProcess at runtime and the result is passed back on the
command line. This way the Qt library and client code is not tainted
by BlueZ's GPL code.

Change-Id: I7961d1b30a5b033dce144d5eb95e687218fceaa7
Reviewed-by: Kai Koehne <kai.koehne@qt.io>