Remove claim we support iOS 5.0
Removed the explicit mentioning of 5.0 and replaced it with a link to
the official list of supported platforms. This reduces the risk of
forgetting left over version indications.

The same treatment was given to the device list. We only mention
the general type of device from now on.

The assumption is that both conditions together, device type and minimum
OS version, will define the list of supported devices.

Change-Id: Ia9f318ad1f1952025e3736589749943193172d86
Reviewed-by: Jake Petroules <jake.petroules@qt.io>