Adapt Qt licensing docs to use the generated license information
Update the licensing overview documentation to list the information
generated by qtattributionsscanner, and drop the locally-maintained
license documentation pages in qtdoc.

The only exception to this is the 3rd-party license for 'Macros for
Building with CMake', which has files across several Qt modules;
list that license in the 'Additional Information' table.

Change-Id: I0e74811012b535e0d1c77ef5fadd78296f44cc86
Reviewed-by: Sami Makkonen <sami.makkonen@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>