Add Qt Cluster Demo
This a a demo of s simple digital instrument cluster.
The demo comes with two skins.
By default the demo will use the so called hybrid skin.
If the argument "sports" is provided the application
will use the sports skin instead.
To be able to preview the custom components in
the Qt Quick Designer the option:
Tools > Options > Qt Quick > Qt Quick Designer >
Use QML emulation layer which is built by the selected

has to be enabled in Qt Creator.

Change-Id: I7d06220062e1fffd6c6849c962bd4f6592d051de
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>