Fix name of ANGLE libraries and section of default graphics backend on Windows
We had to revert the merge of libEGL and libGLESv2 due to BC issues with
the change, so by default we no longer ship QtANGLE but the other 2
libraries.

Additionally we now mention, that the default/packaged version of Qt will
use dynamic opengl and give -opengl dynamic as an example inside the
configure documentation.

This partly reverts 7853e8e51a65aa5c5d9e4c9bb66549e17a2ff022

Change-Id: Icbf61403cd338809204fc36c6e761b03c3acd829
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>