Document CONFIG+=qtquickcompiler as the official way of doing AOT
As 5.11 provides the migration path for what used to be the Qt Quick
Compiler add-on in qtdeclarative itself, we might as well document it to
be the official way of doing ahead-of-time compilation.

The CONFIG+=qmlcache approach turned out to have one grave disadvantage
that was overlooked during the initial release: It ends up working only
inside Qt modules as qml_module.prf is private qmake API.

The CONFIG+=qqc approach on the other hand has been public since its
introduction and it also comes with a CMake integration.

Change-Id: I9bbf4cda1b0a5ff0aa3ac522d77205699255c571
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>