Doc: give property assignments their own line in coding conventions
The point of the convention was that related properties should be part
of the same "block" or "group" of assignments, which can be achieved by
ensuring that they are directly above/below one another. Bunching up
assignments onto one line for a few properties should not be a
recommended practice.

Change-Id: Ie84163b3261e7138d69678e059a91c2c82b43507
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>