Doc: Fix broken external link to Apple Developer page
Change link target so that it points to a location that is less prone
to change in the future. The original document is gone, and the
Apple Developer site has undergone major restructuring in the years
since. By pointing the user to the main landing page, our link will
likely work also in the future, even if Apple decides to redesign their
Apple Developer site.

Task-number: QTBUG-60494
Change-Id: Icff92dcbbe202bc5e2e46ba44eabac610f9b1899
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>