photoviewer: Fix translations
It doesn't make any sense to put .ts files into resources,
because QTranslator can only parse compiled .qm files.
Therefore the application does not load any translations
at all. This error was introduced in 70ce5f4d3ec, when
the example was imported from qtdeclarative.

Anyhow, the previous version of the example in qtdeclarative
required the user to run lrelease manually, which is ugly
and made the example unfit for automatic testing.

Instead, use the new capabilities of qmake to compile
and embed translation files automatically.

Change-Id: Ic02171bfc9994ece3da658b69f0de0a27cb39cc5
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>