Restructure supported platforms
The table has been broken down into individual sections,
as each platform has slightly different requirements and
naming for the information in the table.

Each platform's sub-page also includes the relevant support
information via an include.

Embedded platforms are documented in the Qt for Device
Creation, as the information in the main supported platforms
page was completely out of sync, and the device creations docs
are much better for that domain.

We no longer provide a list of supported platforms for prior
Qt releases. This is information that is redundant, and just
causes confusion when the information is shipped with later
Qt versions. The information is readily available in the
Qt documentation archives online, and should be the
canonical place to look up support information for a given
release.

Change-Id: I5ccf9c7390ccf929d3fe2a7ae88f250cd09cea41
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>