Clarify target platforms of winrt port
Building desktop applications using the winrt port might be possible, but
it is by far not the main reason for having this port. Stating that we
support desktop builds using Qt for WinRT just causes confusion.

Instead we further clarify the term UWP in a short sentence inside the
winrt part of "supported platforms".

Change-Id: I7d1762f5b450dab74ff3c4e1e1aa16586f242319
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>