Doc: Improve CMake code example
Remove the comments, which are mostly superfluous or should
rather be put into the documentation.

Name the project and targets the same.

Guard CMAKE_INCLUDE_CURRENT_DIR with a version check:
This is only needed for CMake 3.6 and older.

Add a resource file to be able to show off CMAKE_AUTORCC.

Remove superfluous helloworld_SRCS variable indirection.

Remove WIN32 argument to add_executable; this seems
superfluous.

Task-number: QTBUG-72159
Change-Id: I39f0d51d23d5dfbbd7a009c0f70c94e5d6296dab
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>