I18N: describe loading translations for Qt modules
Make the first example simpler and move the code for locating
Qt modules to a subsection.

Map Qt modules to catalogs available in the qttranslations
repository.

Explain that devs can either use the meta catalog or pick the
catalogs for the modules that they use (only available way for
new Qt 5 modules).

Task-number: QTBUG-69203
Change-Id: I0727ca3b5547349ab2a1e387040086ead532b688
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>