Update Android documentation
A few things have changed with regards to packaging in Qt 5.14,
so we need the documentation to reflect these changes. One major
change is that we now also support the AAB packaging format,
which also requires multi-arch builds.

When going through the documentation, it was clear that a lot
of the information was already outdated, so I took the opportunity
to clean up a bit. I have removed some stale information, removed
some details that only served to confuse, and rewrote some parts
to match the current reality.

Change-Id: I1d016ea66e18c8f174257defec4af532b4960b15
Reviewed-by: BogDan Vatra <bogdan@kdab.com>
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>