Make docs encourage readers to query locale right
Currently the example shows the usage of "filename" + QLocale::name()
which will not work correctly for users with different locale and
display language. Even the QTranslator::load() docs discourage using
it in this way.

This approach is copied by users of Qt. For example in KDE:
https://github.com/search?q=org%3AKDE+%22%2B+QLocale%3A%3Asystem%28%29.name%28%29%22&type=Code

We should follow the best practice in the first place the user is going
to look at when working with i18n in Qt. Hopefully, with this change
the incorrect code will slowly die out.

Change-Id: I881f50f39740632bbd4640495270aae22bd44d0d
Reviewed-by: Kai Koehne <kai.koehne@qt.io>