Remove Value-Add Modules, Qt Tools from "All Modules" page
Remove the Value-Add Modules section; most of them are not relevant
for Qt 6.0. Qt for Device creation is already listed in
reference-overview.html.

Also remove the Qt Tools section. Up to this day it's mostly a stub (only
mentioning 2 tools). It also competes with the (more complete) list
in reference-overview.html .

Change-Id: I6d86bd20705604596ad0c957b017300aa704365a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Tino Pyssysalo <tino.pyssysalo@qt.io>