Clean up some references to Qt 5 in docs
In several places, "Qt 5" is treated as the brand name,
and should be replaced with just "Qt". The documentation is
part of the repository, so no need to specify the major
release series in the documentation unless we want to refer
to something that changed in a specific version.

Task-number: QTBUG-85207
Change-Id: I6799af72f1d0e06307a724345fb3ef5cbd8ac985
Reviewed-by: Paul Wicking <paul.wicking@qt.io>