Clarify that the CI does not define the reference configurations
For various reasons, not all reference configurations are tested in
the CI, but may instead be tested day to day by developers, and manually
as part of the release process.

By removing the CI references in the documentation we make it clear
that the CI does not define the list of reference configurations.

Change-Id: Ie16dd23afe5cae9d0a67a216cd113ab3c3ca10ec
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>
(cherry picked from commit ce554b6313bac4b760bbced03b113d8cb4803ce3)
Reviewed-by: Tarja Sundqvist <tarja.sundqvist@qt.io>