High-DPI: Extend X11 configuration documentation
Extend the existing "Configuring" section with additional details on
which X11 settings/properties Qt reads, for Qt 5 and Qt 6.

The intention is that someone who is familiar with X11 can understand
which configuration inputs Qt on X11 takes under what conditions, and
then come up with a configuration which works for their use case.

Also mention Ubuntu and Kubuntu as good alternatives for those who would
rather not edit X11 config files.

Pick-to: 6.2
Change-Id: I5fd931aa04374205ca7b6c76b10458eadbe750ff
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>