Doc: Remove entry about CMake incorrect in Qt 6
There is dedicated handling for qml/CMake in Qt 6 and
the CMake API functions that do that are being documented
separately.

The CMake and QML teams are still discussing aspects of
deploying and running qml apps. The result of those
discussions will affect project structure recommendations.

This patch removes incorrect information.

Fixes: QTBUG-92848
Pick-to: 6.2 6.1 6.0
Change-Id: I4bcdac1891721e4efe34f4cf3dd3b25e188e9d88
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>