Doc: Repairs broken link to QWinThumbnailToolButton
A link to `QWinThumbnailToolBarButton` was used when noticing of its
removal in Qt 6.

Nonetheless, Qt 5 does not seem to have any
`QWinThumbnailToolBarButton` class, such that the link was broken.

Instead, a `QWinThumbnailToolButton` (without the `Bar`) exists and is
the actual class that is used for the buttons of
`QWinThumbnailToolBar`.

The link is thus updated to point to `QWinThumbnailToolButton` and its
link-text was modified to mirror the change.

Task-number: QTBUG-96127
Change-Id: I16484ea71eff7099e88730407e4d5fae16daf045
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 1e5197e524e8640f256175400876361918dda197)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>