Doc: Repair broken link about GNOME Icons
`appicon.qdoc` contained a link that aimed to provide more information
about the location and format of icons under the GNOME desktop
environment.

The link is now unavailable and was, hence, replaced with a new link, as
given by Venugopal Shivashankar who added the original link in 2015,
providing information relevant to the context.

The link-text was changed to better represent the content and title of
the new link.

Task-number: QTBUG-96127
Pick-to: 6.2 6.2.0
Change-Id: I36dda8790e146ca02d27d5797b7f56a9b1acddf7
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>