Convert the qmlscene documentation to qml documentation
qmlscene is deprecated.  Much of the same information has been
applicable to the qml runtime tool for a number of years already;
however that wasn't documented until now.

We remove the information about the --dummy-data argument because
it's a means of injecting context properties, which are deprecated.

Task-number: QTBUG-53219
Task-number: QTBUG-96800
Pick-to: 6.2 5.15
Change-Id: I12370d523d43f7ea6ca4ce3ff1ceb8431d87a889
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>