Doc: Update a link to Microsoft's Documentation
The link was to `msdn.microsoft.com` which now redirects to
`docs.microsoft.com`.

While the redirection works, our script to catch broken links on
dev-snapshots builds of the documentation doesn't handle redirection
correctly, reporting it as broken.

Both to appease the broken-links script and to avoid an unneccesary
redirection, the "Side-by-side Assemblies" link in `windows.qdoc` was
modified to point to the new address.

Additionally, the link-text was modified to better represent the content
of the page.

Task-number: QTBUG-96127
Pick-to: 6.2
Change-Id: I436583f8e3fa4aadd1d8e02d3ae0b539c5713610
Reviewed-by: Paul Wicking <paul.wicking@qt.io>