Doc: Remove link to a removed Qt Design Studio example
Qt Design Studio's `Cluster Tutorial` was recently removed.

A link to it was provided in `create-your-first-appliactions` page and
is now removed as well.

The link was registered as an `\externalpage` link in
`external-resources.qdoc` and its entry was removed along the above
mention to it.

Task-number: QTBUG-96127
Change-Id: I11a94523d54c66bd654133178ab6467c7ee3c0ac
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
(cherry picked from commit 23f55b9dcf1202ad5417767c37d7b696951aa66a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>