Clean up section about further information in porting guide
The current links in the 'More Information about Qt 6'
do look a bit arbitrary. It is not clear why 'Qt Creator',
'Qt Overviews' or 'Qt Examples and Tutorials' are particularly
relevant when you're on the 'Porting to Qt 6' page.

Therefore rename the section to 'Further Reading' (about
porting to Qt 6), and remove unrelated links. The section
is not linked to from anywhere else in the documentation,
so just renaming it is safe.

Also, make it clear that the 'Qt 5 and Qt 6 compatibility'
page is for CMake only.

Change-Id: Ib60db0f342607cf60b2f6a4908ed8841541adeb6
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit fd38987ece87506dd3e3757687bf6defa7ca65f1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>