Doc: Update What's new in 6.2 in time for 6.2.1
Certain information was added to the wiki [0] after the creation
of the What's new in 6.2 documentation page, and didn't end up
in the official docs with the 6.2.0 relase. This patch seeks to
remedy that by including the information from the wiki that was
left out.

[0] - https://wiki.qt.io/New_Features_in_Qt_6.2

Fixes: QTBUG-96372
Change-Id: I4bca4e8157e6dcac7638996f9d99219327268c12
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Vladimir Minenko <vladimir.minenko@qt.io>
(cherry picked from commit e0d30d3e1124deba306928f9ada64bc30dc3557e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>