Mention QML type hints in signal/slot documentation
There is a short section about connecting to/from QML. It was a bit
outdated, given that its now possible to use type hints to get a
different type in the meta-object system for QML signals. This change
updates the documentation, and adds a link to the more detailed QML
documentation.

Pick-to: 6.2
Change-Id: I4ccf2fa67d458b77257f1b31598defb13ab1b362
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>