Do not recommend QT_BUILD_TOOLS_WHEN_CROSSCOMPILING
QT_BUILD_TOOLS_WHEN_CROSSCOMPILING was recommended for cross-compiling,
because otherwise qml, qmlscene and qmlpreview were not built for the
device.  Those tools are user-facing tools, and it was a bug that they
were not built for the target.  This is fixed since
qtdeclarative/3e2a85384b1f5454e51a27827114e49746f8837a, and the
recommendation is now invalid.

Pick-to: 6.2
Change-Id: Id1b9962ec1c307d9665e798acbb4be8c4cedfc2f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>