Doc: Add "Building a C++ console application" section
...on the "Getting started with CMake" page.

Split "Build a C++ GUI application" into a "Building a C++ console
application" and, based on that, "Building a C++ GUI application".

Remove the mention of CMAKE_AUTORCC, because we want to discourage its
use.  See QTBUG-87643.  In a subsequent commit, we will add a guide how
to use resources.

Pick-to: 6.3
Task-number: QTBUG-84789
Change-Id: I9434c88538af0d6e7c5e2c86ee86c3d697d7a576
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>